% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{FES}
\alias{FES}
\title{FES method for multiple regression.}
\usage{
FES(x0, y0)
}
\arguments{
\item{x0}{a n*p matrix of explanatory variables}

\item{y0}{a n*q matrix of responses}
}
\value{
Return the estimation of coefficient B.
}
\description{
Factor estimation and selection with the tuning parameter selected algorithm with R.
}
\examples{
\dontrun{
data(samplex)
attach(samplex)
data(sampley)
attach(sampley)
B.hat <- FES(samplex[[1]], sampley[[1]])
}
}
