% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{Gibbs algorithm with Rcpp.}
\usage{
gibbsC(N, a0, b0, n0, mu1, mu2)
}
\arguments{
\item{N}{the number of samples}

\item{a0}{parameter of original distribution}

\item{b0}{parameter of original distribution}

\item{n0}{parameter of original distribution}

\item{mu1}{initial value of 'x' when generating}

\item{mu2}{initial value of 'y' when generating}
}
\value{
Return the generated chain of 'x' and 'y'
}
\description{
To generate two chains 'x' and 'y' using Gibbs algorithm with Rcpp.
}
\examples{
\dontrun{
N <- 5000
b <- 1000
mu1 <- 1
mu2 <- 0.6 
n0 <- 20
a0 <- 7
b0 <- 10
resultC <- gibbsC(N, a0, b0, n0, mu1, mu2)
}
}
