% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{modelerror}
\alias{modelerror}
\title{Model error for N groups.}
\usage{
modelerror(N = 200, x, y, B, sigma, method)
}
\arguments{
\item{N}{number of groups of data "x" and "y"}

\item{x}{a n*p matrix of explanatory variables}

\item{y}{a n*q matrix of explanatory variables}

\item{B}{true coefficient matrix for generating samples}

\item{sigma}{true correlation matrix for generating dataset "x"}

\item{method}{choice of regression method, 1 for FES and 2 for PLS}
}
\value{
Return the estimation of coefficient B.
}
\description{
Mean and standard error of model error for N groups.
}
\examples{
\dontrun{
data(samplex)
attach(samplex)
data(sampley)
attach(sampley)
result <- modelerror(200, samplex, sampley, samplex[[201]], samplex[[202]], method = 1)
}
}
