% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{PLS0}
\alias{PLS0}
\title{PLS method for multiple regression.}
\usage{
PLS0(x0, y0)
}
\arguments{
\item{x0}{a n*p matrix of explanatory variables}

\item{y0}{a n*q matrix of responses}
}
\value{
Return the estimation of coefficient B.
}
\description{
Two-block partial least squares algorithm with R.
}
\examples{
\dontrun{
data(samplex)
attach(samplex)
data(sampley)
attach(sampley)
B.hat <- PLS0(samplex[[1]], sampley[[1]])
}
}
