\name{StatComp21022-package}
\alias{StatComp21022-package}
\alias{StatComp21022}
\docType{package}
\title{
  This package includes homework for this course and research project. 
}
\description{
  This package includes homework for this course and research project. 
}
\author{
Yejing Han
Maintainer: Yejing Han <hanyejing@mail.ustc.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
